Move Projector to utils.py
